# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingCancelRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "BookingCancelRequest",
        "type": "object"
      },
      "BookingCreate": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "integer"
          },
          "service_id": {
            "title": "Service Id",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/BookingSource",
            "default": "api"
          },
          "staff_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Id"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "service_id",
          "starts_at"
        ],
        "title": "BookingCreate",
        "type": "object"
      },
...
      "BookingOverrideCancelRequest": {
        "properties": {
          "reason": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "BookingOverrideCancelRequest",
        "type": "object"
      },
      "BookingOverrideCreateRequest": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "integer"
          },
          "reason": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "service_id": {
            "title": "Service Id",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/BookingSource",
            "default": "api"
          },
          "staff_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Id"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "service_id",
          "starts_at",
          "reason"
        ],
        "title": "BookingOverrideCreateRequest",
        "type": "object"
      },
      "BookingPaymentRead": {
        "properties": {
          "amount_minor_units": {
            "title": "Amount Minor Units",
            "type": "integer"
          },
          "booking_id": {
            "title": "Booking Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "provider_payment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Payment Id"
          },
          "provider_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Session Id"
          },
          "refunded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded At"
          },
          "status": {
            "$ref": "#/components/schemas/BookingPaymentStatus"
          }
        },
        "required": [
          "id",
          "booking_id",
          "provider",
          "provider_session_id",
          "provider_payment_id",
          "amount_minor_units",
          "currency",
          "status",
          "failure_reason",
          "created_at",
          "paid_at",
          "refunded_at"
        ],
        "title": "BookingPaymentRead",
        "type": "object"
      },
      "BookingPaymentRefundRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "BookingPaymentRefundRequest",
        "type": "object"
      },
      "BookingPaymentStatus": {
        "enum": [
          "pending",
          "succeeded",
          "failed",
          "refunded"
        ],
        "title": "BookingPaymentStatus",
        "type": "string"
      },
      "BookingRead": {
        "properties": {
          "business_id": {
            "title": "Business Id",
            "type": "integer"
          },
          "cancel_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Reason"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "service_id": {
            "title": "Service Id",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "staff_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Id"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "business_id",
          "customer_id",
          "service_id",
          "staff_id",
          "starts_at",
          "ends_at",
          "status",
          "source",
          "cancel_reason",
          "created_at"
        ],
        "title": "BookingRead",
        "type": "object"
      },
      "BookingRescheduleRequest": {
        "properties": {
          "new_starts_at": {
            "format": "date-time",
            "title": "New Starts At",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "new_starts_at"
        ],
        "title": "BookingRescheduleRequest",
        "type": "object"
      },
      "BookingSource": {
        "enum": [
          "api",
          "ivr"
        ],
        "title": "BookingSource",
        "type": "string"
      },
      "BookingStatus": {
        "enum": [
          "confirmed",
          "cancelled",
          "pending_payment"
        ],
        "title": "BookingStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "description": "Use the access token returned by login or refresh. Send it as `Authorization: Bearer <access_token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Appointment Voice SaaS API \u2014 multi-tenant booking and availability backend built on a production FastAPI foundation (JWT auth, admin workflows, audit logs, Redis rate limiting, object storage). Salon domain: businesses, staff, services, hours, availability slots, bookings with DB-level double-booking protection. IVR/SMS/calendar: planned.",
    "summary": "Multi-tenant booking and availability backend for local service businesses",
    "title": "Appointment Voice SaaS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/businesses/{business_id}/bookings": {
      "get": {
        "operationId": "list_bookings_endpoint_api_v1_businesses__business_id__bookings_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BookingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "staff_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingRead"
                  },
                  "title": "Response List Bookings Endpoint Api V1 Businesses  Business Id  Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Bookings Endpoint",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "create_booking_endpoint_api_v1_businesses__business_id__bookings_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Booking Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/bookings/override": {
      "post": {
        "description": "Admin override booking creation (P3-012): same conflict rules as a\nnormal create (the DB-level no-overlap constraint still applies for a\nstaff_id with a real conflict \u2014 this does not bypass it), but requires\nan explicit reason and is logged under a distinct audit action so\noverrides are queryable separately from regular bookings.",
        "operationId": "override_create_booking_endpoint_api_v1_businesses__business_id__bookings_override_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingOverrideCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Override Create Booking Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/bookings/{booking_id}": {
      "get": {
        "operationId": "get_booking_endpoint_api_v1_businesses__business_id__bookings__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Booking Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/bookings/{booking_id}/cancel": {
      "post": {
        "operationId": "cancel_booking_endpoint_api_v1_businesses__business_id__bookings__booking_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Booking Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/bookings/{booking_id}/override-cancel": {
      "post": {
        "description": "Admin override cancellation (P3-012): mechanically identical to a\nnormal cancel (frees the slot, offers the waitlist) but requires an\nexplicit reason and is logged under a distinct audit action.",
        "operationId": "override_cancel_booking_endpoint_api_v1_businesses__business_id__bookings__booking_id__override_cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingOverrideCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Override Cancel Booking Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/bookings/{booking_id}/refund": {
      "post": {
        "description": "Manual admin-triggered refund recording (P3-008, ADR 0004 SS6). Does\nnot call any payment-provider API -- records that a refund already\nhappened out-of-band.",
        "operationId": "refund_booking_payment_endpoint_api_v1_businesses__business_id__bookings__booking_id__refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingPaymentRefundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingPaymentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Refund Booking Payment Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/bookings/{booking_id}/reschedule": {
      "post": {
        "operationId": "reschedule_booking_endpoint_api_v1_businesses__business_id__bookings__booking_id__reschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "title": "Booking Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRescheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reschedule Booking Endpoint",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/businesses/{business_id}/clients/{client_id}/bookings": {
      "get": {
        "operationId": "get_client_bookings_endpoint_api_v1_businesses__business_id__clients__client_id__bookings_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingRead"
                  },
                  "title": "Response Get Client Bookings Endpoint Api V1 Businesses  Business Id  Clients  Client Id  Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Client Bookings Endpoint",
        "tags": [
          "clients"
        ]
      }
    },
...
  },
...
}